Upgrade flex and bison.
For C++17 compatibility.
